+++ CHANGES.md
@@ -17,8 +17,11 @@
   - Updated schema for the `num_levels` parameter which now explains the parameter 
     in more detail.
 
-* Improved Extension Demo Panels.
-  See https://github.com/xcube-dev/xcube/issues/1134
+* Improved the demos for the xcube Viewer server-side extensions in various ways (#1134):
+  - enhanced user input validation
+  - added error message label
+  - fixed bugs in code
+  - improved UI styles and general UX
 
 ## Changes in 1.9.0